refactor history command to use command framework
