+++ liquibase-integration-tests/src/test/resources/liquibase/extension/testing/command/history.test.groovy
@@ -13,8 +13,6 @@ Required Args:
   url (String) The JDBC database connection URL
     OBFUSCATED
 Optional Args:
-  changelogFile (String) The root changelog
-    Default: null
   defaultCatalogName (String) The default catalog name to use for the database connection
     Default: null
   defaultSchemaName (String) The default schema name to use for the database connection
