+++ liquibase-standard/src/main/java/liquibase/command/core/HistoryCommand.java
@@ -4,10 +4,14 @@ import liquibase.command.AbstractCommand;
 import liquibase.command.CommandResult;
 import liquibase.command.CommandScope;
 import liquibase.command.CommandValidationErrors;
+import liquibase.command.core.helpers.DbUrlConnectionArgumentsCommandStep;
 import liquibase.database.Database;
 
 import java.io.PrintStream;
 import java.text.DateFormat;
+
+import static liquibase.command.core.HistoryCommandStep.DATE_FORMAT_ARG;
+import static liquibase.command.core.HistoryCommandStep.FORMAT_ARG;
 
 /**
  * @deprecated Implement commands with {@link liquibase.command.CommandStep} and call them with {@link liquibase.command.CommandFactory#getCommandDefinition(String...)}.
@@ -51,12 +55,12 @@ public class HistoryCommand extends AbstractCommand {
 
     @Override
     public CommandResult run() throws Exception {
-        final CommandScope commandScope = new CommandScope("internalHistory");
+        final CommandScope commandScope = new CommandScope(HistoryCommandStep.COMMAND_NAME);
         commandScope.setOutput(getOutputStream());
 
-        commandScope.addArgumentValue(InternalHistoryCommandStep.DATABASE_ARG, this.getDatabase());
-        commandScope.addArgumentValue(InternalHistoryCommandStep.DATE_FORMAT_ARG, this.dateFormat);
-        commandScope.addArgumentValue(InternalHistoryCommandStep.FORMAT_ARG, this.format);
+        commandScope.addArgumentValue(DbUrlConnectionArgumentsCommandStep.DATABASE_ARG, this.getDatabase());
+        commandScope.addArgumentValue(DATE_FORMAT_ARG, this.dateFormat);
+        commandScope.addArgumentValue(FORMAT_ARG, this.format);
 
         commandScope.execute();
