+++ liquibase-standard/src/main/resources/META-INF/services/liquibase.command.CommandStep
@@ -14,7 +14,6 @@ liquibase.command.core.FutureRollbackFromTagSqlCommandStep
 liquibase.command.core.FutureRollbackSqlCommandStep
 liquibase.command.core.GenerateChangelogCommandStep
 liquibase.command.core.HistoryCommandStep
-liquibase.command.core.InternalHistoryCommandStep
 liquibase.command.core.InternalSnapshotCommandStep
 liquibase.command.core.ListLocksCommandStep
 liquibase.command.core.MarkNextChangesetRanCommandStep
