+++ liquibase-standard/src/test/groovy/liquibase/command/CommandResultsTest.groovy
@@ -1,6 +1,6 @@
 package liquibase.command
 
-import liquibase.command.core.InternalHistoryCommandStep
+import liquibase.command.core.HistoryCommandStep
 import liquibase.util.StringUtil
 import spock.lang.Specification
 
@@ -9,9 +9,9 @@ class CommandResultsTest extends Specification {
     def "can get and set value"() {
         when:
         def commandResults = new CommandResults(new TreeMap<>([
-                "a": "value from a",
-                "bool": true,
-                (InternalHistoryCommandStep.DEPLOYMENTS_RESULT.getName()): new InternalHistoryCommandStep.DeploymentHistory()
+                "a"                                              : "value from a",
+                "bool"                                           : true,
+                (HistoryCommandStep.DEPLOYMENTS_RESULT.getName()): new HistoryCommandStep.DeploymentHistory()
         ]), new CommandScope("history"))
 
         then:
@@ -20,7 +20,7 @@ class CommandResultsTest extends Specification {
         commandResults.getResult("bool") == true
         commandResults.getResult("invalid") == null
 
-        commandResults.getResult(InternalHistoryCommandStep.DEPLOYMENTS_RESULT).toString() == "0 past deployments"
+        commandResults.getResult(HistoryCommandStep.DEPLOYMENTS_RESULT).toString() == "0 past deployments"
 
         StringUtil.join(commandResults.getResults(), ", ") == "a=value from a, bool=true, deployments=0 past deployments"
     }
